Moving to compartmentalizing the Energy System
